titan-poss-poss-web-f78a2c9ef984/titan-poss-poss-web-f78a2c9ef984/libs/eposs/grn-request-approvals/data-access-grn-request-approvals/src/lib/+state/grn-request-approvals.selectors.ts
import { createSelector } from '@ngrx/store';
import { selectGrnRequestApprovalState } from './grn-request-approvals.reducer';
import { grnRequestListSelector } from './grn-request-approvals.entity';

export const grnRequestApprovalList = createSelector(
  selectGrnRequestApprovalState,
  state => state.grnRequestList
);

const selectGrnRequestApprovalList = createSelector(
  grnRequestApprovalList,
  grnRequestListSelector.selectAll
);

const selectIsLoading = createSelector(
  selectGrnRequestApprovalState,
  state => state.isLoading
);
const selectError = createSelector(
  selectGrnRequestApprovalState,
  state => state.error
);

const selectHasSaved = createSelector(
  selectGrnRequestApprovalState,
  state => state.hasSaved
);

const selectHasUpdated = createSelector(
  selectGrnRequestApprovalState,
  state => state.hasUpdated
);
export const grnRequestApprovalSelector = {
  selectGrnRequestApprovalList,
  selectIsLoading,
  selectError,
  selectHasSaved,
  selectHasUpdated
};
